# projects/go/cellphone/api/openapi.yaml
openapi: 3.1.0

info:
  title: Cellphone API
  description: An API that allows consumers to get new and unused cellphones from different providers
  version: 2.0.0
servers:
  - url: http://localhost:8080/
    description: Local development server
tags:
  - name: cellphone
  - name: provider
paths:
  /Provider/{id}:
    get:
      summary: Lookup provider by id
      description: Lookup provider by id, which is the primary key.
      tags:
        - provider
      parameters:
        - name: id
          in: path
          required: true
          description: Provider ID
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Provider was found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        404:
          description: Provider was not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        500:
          description: An unexpected error occurred.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /Provider/{id}/count:
    get:
      summary: How many phones a provider has
      description: Gets how many phones a provider has by its ID
      tags:
        - provider
      parameters:
        - name: id
          in: path
          required: true
          description: Provider ID
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Provider was found and the given count returned
          content:
            appplication/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    format: int64
                    example: 64
                    required: true
                    description: How many phones exist for the given provider
        400:
          description: Provider ID not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        500:
          description: An unexpected error occurred
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /Provider/ByName/{name}:
    get:
      summary: Get provider by name
      description: Gets a provider by its Name
      tags:
        - provider
  /Provider:
    delete:
      summary: Deletes an existing Provider
      tags:
        - provider
      parameters:
      responses:
    post:
      summary: Creates a new Provider
      tags:
        - provider
      parameters:
      responses:
    patch:
      summary: Updates a new Provider
      tags:
        - provider
      parameters:
      responses:
  /Cellphone/{id}:
    get:
      summary: Gets a phone by ID
      description: Gets a cellphone by its ID
      tags:
        - cellphone
      parameters:
        - name: id
          in: path
          required: true
          description: Cellphone ID
          schema:
            type: integer
            format: int64
      responses:
    post:
      summary: Inserts multiple cellphones
      description: Inserts multiple cellphones into the database for a given Provider ID
      tags:
        - cellphone
      parameters:
        - name: id
          in: path
          required: true
          description: Provider ID
      responses:
  /Cellphone:
    post:
      summary: Fetch cellphone
      description: Retrieves an existing cellphone and deletes it from the database
      tags:
        - cellphone
      parameters:
      responses:
components:
  schemas:
    ApiError:
      type: object
      required:
        - error
      properties:
        error:
          type: string
    Provider:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 88909
          description: Table unique identifier
        name:
          type: string
          example: "Vivo"
          description: Provider's name
        total:
          type: integer
          format: int64
          example: 12000
          description: Total phone numbers stored for this provider
      required:
        - id
        - name
        - total
    Cellphone:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 22528
          description: Cellphone's ID
        provider_id:
          type: integer
          format: int64
          description: The provider that owns this number
        number:
          type: string
          example: "999887766"
          description: The phone number
      required:
        - id
        - provider_id
        - total